fix interface IOptions
