+++ package/src/index.d.ts
@@ -1,28 +1,7 @@
 import * as T from './types';
 
-export interface IOptions {
-	input?: boolean;
-	type?: T.TypesCalendar;
-	months?: number;
-	jumpMonths?: number;
-	date?: Partial<T.IDate>;
-	settings?: Partial<{
-		lang: string;
-		iso8601: boolean;
-		range: Partial<T.IRange>;
-		selection: Partial<T.ISelection>;
-		selected: Partial<T.ISelected>;
-		visibility: Partial<T.IVisibility>;
-	}>;
-	locale?: Partial<T.ILocale>;
-	actions?: Partial<T.IActions>;
-	popups?: T.IPopups;
-	CSSClasses?: Partial<T.ICSSClasses>;
-	DOMTemplates?: Partial<T.IDOMTemplates>;
-}
-
 export default class VanillaCalendar {
-	constructor(selector: HTMLElement | string, options?: Partial<IOptions>);
+	constructor(selector: HTMLElement | string, options?: Partial<T.IOptions>);
 
 	input: boolean;
 	type: T.TypesCalendar;
