+++ package/src/vanilla-calendar.ts
@@ -5,29 +5,8 @@ import update from '@scripts/update';
 import init from '@scripts/init';
 import destroy from '@scripts/destroy';
 
-export interface IOptions {
-	input?: boolean;
-	type?: T.TypesCalendar;
-	months?: number;
-	jumpMonths?: number;
-	date?: Partial<T.IDate>;
-	settings?: Partial<{
-		lang: string;
-		iso8601: boolean;
-		range: Partial<T.IRange>;
-		selection: Partial<T.ISelection>;
-		selected: Partial<T.ISelected>;
-		visibility: Partial<T.IVisibility>;
-	}>;
-	locale?: Partial<T.ILocale>;
-	actions?: Partial<T.IActions>;
-	popups?: T.IPopups;
-	CSSClasses?: Partial<T.ICSSClasses>;
-	DOMTemplates?: Partial<T.IDOMTemplates>;
-}
-
 export default class VanillaCalendar extends DefaultOptionsCalendar implements T.IVanillaCalendar {
-	constructor(selector: HTMLElement | string, options?: Partial<IOptions>) {
+	constructor(selector: HTMLElement | string, options?: Partial<T.IOptions>) {
 		super();
 
 		this.HTMLElement = (typeof selector === 'string' ? document.querySelector(selector) : selector) as HTMLElement;
